fix(): `endCurrentTransform` `mouseUp` overlap
