+++ src/shapes/Object/InteractiveObject.ts
@@ -120,8 +120,7 @@ export class InteractiveFabricObject<
   _updateCacheCanvas() {
     const targetCanvas = this.canvas;
     if (this.noScaleCache && targetCanvas && targetCanvas._currentTransform) {
-      const target = targetCanvas._currentTransform.target,
-        action = targetCanvas._currentTransform.action;
+      const { target, action } = targetCanvas._currentTransform;
       if (
         this === (target as InteractiveFabricObject) &&
         action.startsWith('scale')
@@ -136,20 +135,20 @@ export class InteractiveFabricObject<
    * Determines which corner is under the mouse cursor, represented by `pointer`.
    * This function is return a corner only if the object is the active one.
    * This is done to avoid selecting corner of non active object and activating transformations
-   * rather than drag action. The default behaviour of fabricJS is that if you want to trasform
+   * rather than drag action. The default behavior of fabricJS is that if you want to transform
    * an object, first you select it to show the control set
    * @private
    * @param {Object} pointer The pointer indicating the mouse position
    * @param {boolean} forTouch indicates if we are looking for interaction area with a touch action
-   * @return {String|Boolean} corner code (tl, tr, bl, br, etc.), or 0 if nothing is found.
+   * @return {String} corner code (tl, tr, bl, br, etc.), or an empty string if nothing is found.
    */
-  _findTargetCorner(pointer: Point, forTouch = false): 0 | string {
+  _findTargetCorner(pointer: Point, forTouch = false): string {
     if (
       !this.hasControls ||
       !this.canvas ||
       (this.canvas._activeObject as InteractiveFabricObject) !== this
     ) {
-      return 0;
+      return '';
     }
 
     this.__corner = undefined;
@@ -182,7 +181,7 @@ export class InteractiveFabricObject<
       // this.canvas.contextTop.fillRect(lines.rightline.d.x, lines.rightline.d.y, 2, 2);
       // this.canvas.contextTop.fillRect(lines.rightline.o.x, lines.rightline.o.y, 2, 2);
     }
-    return 0;
+    return '';
   }
 
   /**
